refactor(PrimitivesManager): move its location from plb.engine.primitive to plb.engine for its dependence on controllers
