+++ plb/engine/controller/robot_controller.py
@@ -10,7 +10,7 @@ from plb.config.utils import make_cls_config
 from .controller import Controller
 from plb.urdfpy import DiffRobot, Robot, Collision, DEVICE
 from plb.engine.primitive.primitive import Box, Sphere, Cylinder, Primitive
-from .primitive_controller import PrimitivesController
+
 
 ROBOT_LINK_DOF = 7
 ROBOT_LINK_DOF_SCALE = tuple((0.01 for _ in range(ROBOT_LINK_DOF)))
